+++ packages/codebytes/src/types.ts
@@ -6,12 +6,22 @@ import { LanguageOption } from './consts';
 export interface CodebytesChangeHandler {
   (text: string, language: LanguageOption): void;
 }
-export interface CodeByteEditorProps
-  extends Omit<BackgroundProps, 'onCopy' | 'bg'> {
+
+export interface CodebytesCopyFormatterParams {
+  text: string;
+  language: LanguageOption;
+}
+
+export type CodebytesCopyFormatter = ({
+  text,
+  language,
+}: CodebytesCopyFormatterParams) => string;
+
+export interface CodeByteEditorProps extends Omit<BackgroundProps, 'bg'> {
   text?: string;
   language?: LanguageOption;
   hideCopyButton?: boolean;
-  onCopy?: CodebytesChangeHandler;
+  copyFormatter?: CodebytesCopyFormatter;
   snippetsBaseUrl?: string;
   onEdit?: CodebytesChangeHandler;
   onLanguageChange?: CodebytesChangeHandler;