+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.22.0](https://github.com/Codecademy/client-modules/compare/@codecademy/tracking@0.21.1...@codecademy/tracking@0.22.0) (2022-05-16)
+
+
+### Features
+
+* **tracking:** Adding misc field for click and visit events ([57ae584](https://github.com/Codecademy/client-modules/commit/57ae584d3519db2d40b9a4c021fbee309df59d75))
+
+
+
+### [0.21.1](https://github.com/Codecademy/client-modules/compare/@codecademy/tracking@0.21.0...@codecademy/tracking@0.21.1) (2022-05-13)
+
+
+### Bug Fixes
+
+* **tracking:** pick content_ids ([bd8b0a6](https://github.com/Codecademy/client-modules/commit/bd8b0a63eb6d8cfda2c0471fddcb7954898fbc75))
+
+
+
+## [0.21.0](https://github.com/Codecademy/client-modules/compare/@codecademy/tracking@0.20.0...@codecademy/tracking@0.21.0) (2022-05-13)
+
+
+### Features
+
+* **tracking:** add UserSharedData to user impressions ([6378ec2](https://github.com/Codecademy/client-modules/commit/6378ec2cb4f076210fa842dcfb07b43717afd1f9))
+
+
+
 ## [0.20.0](https://github.com/Codecademy/client-modules/compare/@codecademy/tracking@0.19.0...@codecademy/tracking@0.20.0) (2022-04-11)